Added myself and fixed typos
I added myself to the census and fixed two small formatting typos where an extra bullet point was displayed on one of the names and there was a trailing "/" on one of the URLs. 
Said URL was not displaying properly as a link on the live page and not displaying at all on the main repo page, maybe this will fix it.